+++ CHANGELOG.md
@@ -7,6 +7,7 @@ Breaking changes:
 New functionality added in a backwards-compatible manner:
 * [OLMIS-5625](https://openlmis.atlassian.net/browse/OLMIS-5625): Added OpenlmisValidator class.
 * [OLMIS-5625](https://openlmis.atlassian.net/browse/OLMIS-5625): Added OpenlmisArrayDecorator class.
+* [OLMIS-5356](https://openlmis.atlassian.net/browse/OLMIS-5356): Extended OpenlmisResource with the getAll method.
 
 Improvements:
 * [OLMIS-3696](https://openlmis.atlassian.net/browse/OLMIS-3696): Added dependency and development dependency locking.